+++ src/utils.ts
@@ -1,10 +1,11 @@
+import { asArray, asString } from "ol/color";
 import { buffer } from "ol/extent";
 import Geometry from "ol/geom/Geometry";
 import Map from "ol/Map";
 import { Vector } from "ol/source";
 import { getArea } from "ol/sphere";
 
-export type AreaUnitEnum =  "m2" | "ha";
+export type AreaUnitEnum = "m2" | "ha";
 
 /**
  * Calculate & format the area of a polygon
@@ -42,4 +43,15 @@ export function fitToData(
 ) {
   const extent = olSource.getExtent();
   return olMap.getView().fit(buffer(extent, bufferValue));
+}
+
+/**
+ * Translate a hex color to an rgba string with opacity
+ * @param hexColor - a hex color string
+ * @param alpha - a decimal to represent opacity
+ * @returns - a 'rgba(r,g,b,a)' string
+ */
+export function hexToRgba(hexColor: string, alpha: number) {
+  const [r, g, b] = Array.from(asArray(hexColor));
+  return asString([r, g, b, alpha]);
 }